refactor: ViewCountByUUID 타입 정의
